cmd: add flux-uri
Problem: A command line interface for resolving Flux jobids and
uri-resolver URIs to job URIs is not currently available.

Add flux-uri.py for this purpose. The command can take a jobid
or URI in any supported resolver scheme and will return a
FLUX_URI for the target instance.

For convenience, --local and --remote options are provided to
attempt to convert a resolved job URI to its local or remote form.