testsuite: add tests for FluxURIResolver, flux-uri
Problem: No tests exist for the Python FluxURIResolver class and
its front-end tool flux-uri.

Add a small set of tests for the base URI and FluxJobURI classes
to the python tests as t/python/t0025-uri.py, and a larger set of
functionality tests in t2802-uri-cmd.t which use the flux-uri(1)
command.