libutil: add uri_resolve()
Problem: Multiple flux-core utilities could benefit from calling
out to flux-uri to resolve high-level URIs to native URIs, but this
requires a lot of boilerplate code to launch the external process
and read its output, etc.

Add a uri_resolve() convenience function to libutil which takes a
possibly unresolved URI and calls `flux uri` on it, returning the
resolved URI on success.