doc: add domainrefs sphinx plugin
Problem: It is cumbersome to cross reference manpages with sphinx,
when it should be simple.

Add the `domainrefs.py` sphinx plugin from

 https://github.com/mitogen-hq/mitogen/blob/master/docs/domainrefs.py

which allows simple addition of multiple domain cross-references to
the sphinx conf.py.

Now referencing another manpage is as simple as :man1:`flux` for example.

Extend PYTHONPATH in sphinx commands so that sphinx can find extension
in srcdir.