ci: move codecov into docker
Problem: For centos8, the version of gcov outside the container
is not the same as that inside the container, which causes errors
during coverage collection due to differences in the gcda output
file formats.

Run the codecov collection tool inside the docker container instead
of outside for the system checks that run in centos8.