flux-top: resolve argument with uri_resolve()
Problem: The argument to flux-top(1) (if provided) can only be a
valid jobid in the enclosing instance, but for enhanced usability
any high-level URI accepted by `flux uri` should be allowed.

If an argument is provided to `flux top`, attempt to resolve it with
the new `uri_resolve()` function, which calls out to `flux uri`.
Thus, any high-level URI accepted by `flux uri` can also be used
directly with `flux top`.