python: raise error if job not running in flux.uri.resolvers.jobid
Problem: The "jobid" URI resolver returns a URI for Flux instances that
have already completed, in most cases resulting in a less useful later
error when the caller attempts to connecto to the completed instance.

Update the jobid URI resolver to raise an error if the target jobid
is not running.