flux-top: refactor flux-top initialization
Problem: The top context object `struct top` is allocated on the
stack in main() and created destroyed in parts in the same function.
It would be better to use the more typical create/destroy interface
for this data structure to encourage reusability.

Add top_create(), top_destroy() and top_run() so that `struct top`
has a more typical object interface.

This has the drawback that some checks which were done before screen
initialization are now done after. This means that some errors may
occur after the screen is blanked, so a user may see an annoying
blink effect. Since errors are still visible after the screen is
restored, this is only a minor annoyance and can be fixed at a
later time.