flux-top: allow top to call itself recursively
Problem: When viewing jobs in flux-top(1), it would be nice to select
a job which is an instance of Flux, and recursively call top on that
job without having to exit the top session.

Add the ability to select a job by moving through the joblist pane
with either 'j' and 'k' or the arrow keys. While a job is selected,
hitting enter will recursively call top on that job if the job is
a Flux instance. When run recursively, the 'q' key will "pop back"
to the parent job. Ctrl-c will quit flux-top immediately.

Jobs that are Flux instances are now drawn in blue to assist in
their identification.