t/scripts/runpty.py: add --stderr=FILE argument
Problem: Sometimes we want to run a command under a pty but capture
stderr in a different file, e.g. to check for expected errors, but
the runpty.py test script doesn't support this.

Add a simple `--stderr=FILE` option to runtpy.py to redirect stderr
to a file before launching the child process.