t2801-top-cmd.t: run all tests under a tty
Problem: It is slightly awkward in flux-top to ensure some or most
errors are discovered and printed before curses and initialized so
that tests for these errors can be run without a tty.

Now that runtpy.py can redirect stderr to a file, run most flux-top
tests in t2801-top-cmd.t with a tty, freeing up the implementation to
initialize the curses screen whenever it needs to do so.