flux-uri: Set default for --local from FLUX_URI_RESOLVE_LOCAL
Problem: In testing it is useful to resolve all native URIs in their
local form, but this currently requires passing `--local` to `flux uri`
or awkwardly appending the `?local` query parameter to URIs to resolve.

Set the default for `--local` from the FLUX_URI_RESOLVE_LOCAL
environment variable, which could be set during testing and development
to force all URIs to their local form.