t/scripts/runpty.py: add --input, --no-output options
Problem: There is no way to send input to test programs run under
the runpty.py test command. Also, it would be useful in some cases
to easily redirect output to /dev/null instead of a file or screen.

Add a --input=FILE option that takes input in asciicast v2 format
to send to the program under runpty's control. The offset timestamps
in the asciicast file will be used to schedule the writes to the child
process' pty.

Add a --no-output option which redirects ouput to /dev/null.