flux-top: check that terminal has required capabilities
Problem: When flux-top is run with TERM set to a terminal without
necessary capabilites (e.g. TERM=dumb), it displays a blank screen
with cursor in lower right corner, which is not optimal.

Check for a basic terminal capability "cup" (cursor addressing), and
if this fails abort flux-top with "terminal does not support required
capabilities"