t/scripts/runpty.py: set TERM=xterm by default
Problem: The runpty.py script inherits the TERM environment variable
from the current environment, but this may not be appropriate for
running a test terminal application, e.g. if TERM=dumb.

Set TERM to a reasonable default "xterm" for the child process.
Add a `--term=TERMINAL` option so that this value can be overidden.