libterminus: disconnect pty client on remote error
Problem: When a pty client gets an error response from a write
to a pty service, an error is logged but no other action is taken.
If the remote server went away, then the client will be stuck
in the pty since even the detach command can't be processed.

In the case of a write error, assume there is no recovery and
immediately detach from the pty with an error message ("pty disappeared"
for ENOSYS, generic error otherwise)

Fixes #4005