doc: update domain roles to search manpages
Problem: When adding cross references to other documents, only the
primary page can be linked, not a secondary page.  This can lead
to broken links in the documentation.

Solution: Update to add roles that will search manpage listings
for the correct primary page to link against.  Add a flux-undocumented
placeholder page in the event any references are undocumented
in the future.