cmd/flux-jobs: update -a and -A options
Problem: The -A function lists jobs of all users and jobs in all
job states.  This option will override the --filter option if
specified by the user.

The language of the -a option is confusing.  It states that it
lists all jobs for the current user, but the current user is
the default.

Solution: Re-work the -A option to only list jobs for all users
(i.e. --user=all).  This allows the -A option to work with the
--filter option.

Update the language of the -a option to list jobs in all job states.

Subsequently, it may be common for users to run "flux jobs -a -A" to
list jobs in all states for all users.

Update documentation appropriately.

Update tests as needed.