flux-top: make title path-like string of jobids
Problem: It would be useful in flux-top to have an indication of whether
the information being displayed is from a recursive call to top, and if
so, the depth of the recursion. This can be a good hint to the user what
will happen when they press the 'q' key, for example.

Build the top title as a path-like string of jobids, starting with a
"/" for the root instance, or the jobid if `flux top` was called on
a specific job. This allows users to easily see the "path" of jobids
they've traversed in the current flux-top session, and gives a useful
hint for how many times 'q' needs to be used to pop back up to the
original jobid.