flux-top: try to continue if RPCs return ENOSYS
Problem: When running top on a job that is still starting, the
initial RPCs to the scheduler and job-list modules may return
ENOSYS. This causes flux-top to issue a fatal error an exit when
a simple retry would do.

Simply continue when getting ENOSYS errors from RPCs. Other errors
will still cause flux-top to exit immediately.