libutil/uri: fix segfault in uri_resolve()
Problem: If the string passed to uri_resolve() begins with a "/",
then then yuarel_parse() returns success, but with scheme=NULL. This
causes a segfault since the function doesn't check for a valid scheme
before using strcmp() to check for known "native" Flux URIs.

Only test for "native" Flux URIs if there is a scheme to compare.