flux-top: split inactive into completed/failed
Problem: It may be useful when browsing jobs with flux-top to see
the number of inactive jobs that didn't complete successfully vs
the number that did.

Split the inactive job count into "completed" vs "failed", where
completed are jobs that ran to completions and exited with a zero
status, and failed jobs are those that exited with non-zero status,
were canceled, or timed out.