flux-proxy: add --reconnect option
Problem: flux-proxy fails if connection to broker is lost, but
it might be useful to have an option to try to reconnect.

Register a flux comms error callback.  If the callback is invoked
and --reconnect was provided, then attempt to reconnect, retrying
forever or until the user kills flux-proxy.

Remain in the reconnect loop until the broker enters RUN state.
so that reconnect occurs after flux services are initialized,
avoiding unexpected errors for flux-proxy(1) users.

Call router_renew() from the handle error callback upon reconnect
so that clients of flux-proxy don't lose their event subscriptions
and service registrations.