python: support FLUX_URI_RESOLVE_LOCAL in jobid URI resolver
Problem: When a hierarchical jobid is resolved via the jobid URI
resolver, the FLUX_URI_RESOLVE_LOCAL environment variable is
not used. This can cause hangs during testing while the ssh
connector prompts for a password.

Set the force_local flag to True if FLUX_URI_RESOLVE_LOCAL is set
in the current environment.