python: add uri and uri.local attributes to JobInfo
Problem: Currently, the non-obvious job.user.uri JobInfo attribute
holds the URI submitted as a job memo by a Flux instance running as a
job, which means this attribute is also used to indicate if a job is
a running Flux instance. It would be better to promote this important
value to a top-level job attribute.

Add a new job.uri property to JobInfo objects. This property returns
a flux.uri.JobURI if the job has a job.user.uri attribute, otherwise
returns None.

Allow format fields {uri} and {uri.local} to be used with flux-jobs.