python: obey FLUX_URI_RESOLVE_LOCAL for JobURI
Problem: When returning the string representation of a flux.uri.JobURI
the original URI string is returned, but it may be convenient to
return the local URI when FLUX_URI_RESOLVE_LOCAL is set.

When instantiating the JobURI class, set the class variable force_local
to True if the FLUX_URI_RESOLVE_LOCAL environment variable is
non-empty.

Then, in the JobURI str() method, return the local URI if force_local
is True.