flux-jobs: add -R, --recursive option
Problem: There is no way to easily list jobs recursively in a FLux
instance.

Add a -R, --recursive option to flux-jobs(1) which will recursively
list jobs that are Flux instances up to an optional level given by
an -L, --level=N option.

Similar to ls -R, the relative "path" to each job is printed before
the listing of jobs. If the --stats option is used, then stats for
each job is also included.