flux-jobs: color instances in blue
Problem: Flux instances are not denoted or flagged in the output
of flux-jobs(1). Since the hierarchical nature of Flux is integral
to its design and usage, child jobs should be easily distinguishable
in the job listing output.

Follow the example of flux-top(1) and color running Flux instances
in blue. This gives users a clear indication of which and how many
jobs in the current listing are instances that may be targets of
further inquiry.