librouter: add router_renew()
Problem: now that a router can reconnect to a rebooted
broker, we need a way for the router to re-establish event
subscriptions and service registrations that it maintains
on behalf of its clients.

Add router_renew() which synchronously reestablishes all of
its cached subscriptions and service registrations.  This
can optionally be called after a successful flux_reconnect(3).