python: add EmptyObject in place of empty AnnotationsInfo
Problem: Generation of an "empty" object for formatting possibly missing
attributes with flux-jobs or other tools is more general than just for
AnnotationsInfo objects, but the AnnotationsInfo class is the only class
which supports this feature currently.

Add an EmptyObject class spefically for the purpose of allowing formatting
of missing data. The EmptyObject class can be inserted in place of a
real class and will return suitable "empty" data for various conversions
(str(), repr(), int() and float()), and also intercepts calls to .format()
method and returns an empty string with the current format spec.