python: support instance-specific data for JobInfo
Problem: Instance-specific data for jobs that are also instances of Flux
is not available from the JobInfo interface nor flux-jobs(1).

Add a JobInfo.get_instance_info() method which attempts to fetch instance
resource and job stats for jobs which have a valid URI and are also in a
"running" state.

Add a new InstanceInfo class which provides instance specific data under a
new JobInfo .instance attribute. If a job is not a Flux instance, or the
fetch of instance data fails (e.g. if the instance is still starting or
shutting down), an empty instance object is used such that `instance.*`
will return empty values.

Include the following attributes in the InstanceInfo class for now:

 - instance.resources - access the SchedResourceList object for the instance
 - instance.stats - access a StatsInfo object for the instance, which wraps
    a JobStats object, but provides a reasonable default str() and repr()
    representation.
 - instance.utilization - return a percentage of cores currently in use by
    jobs in the instance
 - instance.gpu_utilization - same as above, but for GPUs, if present
 - instance.progress - return a percentage of jobs which have completed