testsuite: update t2800-jobs-recursive.t
Problem: t2800-jobs-recursive.t doesn't test handling of jobs owned
by users other than the current user and the --recurse-all option.

Add support for running a fake job as an alternate user to this test,
and ensure that this job is skipped by default for recursion, even
with -A, and recursion on this job is attempted wtih --recurse-all.