cmd: add flux-pstree.py
Problem: Flux doesn't have a tool to effectively display and list
information for job hierarchies.

Add a `flux pstree` command, which by default displays a tree of
running jobs by job name, similar to the pstree(1) system command.

Like pstree(1), identical leaves of the tree are combined, which
should result in condensed output when many jobs share the same name.

The flux-pstree(1) command supports custom node labels, which accept
any `JobInfo` format fields (i.e. anything supported by flux-jobs(1)),
and "parent" nodes (i.e. Flux instances) may be labeled separately
from simple jobs.

The command only lists actively running jobs by default, but similar
to flux-jobs(1) supports an `-a, --all` option to query jobs in all
states for the current user. In the case `-a` is used, then the label
will automatically include the job state, unless separately overridden
on the commandline.

flux-pstree(1) additionally supports listing of extended information
before the tree display with the `-x, --extended` and `-d,
--details=NAME` options, e.g.:

 $ flux pstree -x
       JOBID USER     ST NTASKS NNODES  RUNTIME
    ƒ3KrufpK grondo    R      1      1   13.61s flux
     ƒUXhfWP grondo    R      1      1   11.77s └── flux
     ƒUAwqZZ grondo    R      1      1   10.03s     └── flux
    ƒ2DrcgAs grondo    R      1      1   6.400s         └── sleep

Several builtin detail formats are available via the `-d, --details`
option, e.g. progress, resources, and stats.

 $ flux pstree --details=stats
       JOBID           STATS              RUNTIME
    ƒ3KrufpK     PD:2 R:1 CD:0 F:0        0:01:24 flux
     ƒUXhfWP     PD:0 R:1 CD:0 F:0        0:01:22 └── flux
     ƒUAwqZZ   PD:934 R:1 CD:65 F:0       0:01:20     └── flux
    ƒ2LrqHPN                              0:00:01         └── sleep

A custom details format can be supplied with the `--detail-format=FORMAT`
option.

Trees are truncated if they exceed the current value of the COLUMNS
environment variable, or the terminal size if COLUMNS is not set. To
disable this behavior, a `-l, --long` option is provided.

By default, the enclosing instance, or root of the tree, is skipped
when -x, --extended and/or -d, --details=NAME options are used, but
the root is displayed when flux-pstree(1) is run without these options.
This behavior can be changed with `--skip-root=[yes|no]`.

See `flux pstree --help` for other available options.