broker: avoid inappropriate quorum.timeout
Problem: spurious 'quorum delayed' log messages are emitted on rank 0
of a system instance.

The timer is armed even if the timeout value is set to -1, as it the
case for the system instance, where the systemd unit file sets th
quorum.timeout attribute to "none".

Don't arm the timer unless the timeout is greater than zero.

Fixes #4017

Co-authored-by: Mark A. Grondona <mark.grondona@gmail.com>