python: update !P format character to 5 character width
Problem: The limit of 4 characters for !P conversion spec causes
issues in boundary conditions such as 0.05% and 99.9%.

Update the conversion to allow one digit more precision. This causes
the field to now take 5 characters instead of 4. As a side benefit,
CORE% can now be used as the default heading instead of CPU%, which
could be seen as slightly misleading.

Update documentation.

Update tests.