libutil: fix environment_push() with multiple entries
Problem: When environment_push() is used to push a colon-separated
path onto the front of an environment variable, it does so in the
wrong order. This is because the path is traversed in order and each
entry added onto the front individually. This reverses the order of
the elements.

When pushing a split path with before=true, reverse the argz elements
after they're split so they are pushed in the correct order.

Fixes #4038