conf: put builddir before srcdir for intree PYTHONPATH
Problem: For unknown reasons, when srcdir comes before builddir in the
intree PYTHONPATH, and the build is vpath or distcheck build, python
fails to find the `_core._flux` module.  (This was found with python
3.6 at least) This recently became a problem when a bug that reversed
the order of paths prepended to environment variables was fixed.

Since the order of builddir vs srcdir should not matter for PYTHONPATH,
reverse the order in libflux/Makefile.am:INTREE_PYTHON_PATH definition,
so we use the same order as before the prepend fix.