testsuite: do not run t9000-system.t with debug
Problem: The t9000-system.t test is run whenever debug is set in
the testsuite, but this test script is only appropriate when testing
in a system instance.

Only run this test when FLUX_ENABLE_SYSTEM_TESTS is set to avoid
any attempts to run it when debug is set for other reasons.