resource: drain nodes in broker.torpid group
Problem: jobs continue to be scheduled on unresponsive nodes.

Monitor the broker.torpid group from the resource module, and drain
any nodes that enter the group, if they are not already drained.
When a node becomes responsive again, it is not automatically undrained.
A manual "flux resource undrain" is required to make a node eligible
for scheduling after it has been drained.

Fixes #3448