build: reverse order of libflux-internal and libflux-core
Problem: Placing libflux-internal after libflux-core in LIBADD
and LDADD appears to confuse some versions of libtool into adding
an rpath to the eventual install location of libflux-core.so. This
causes unit tests, even when using the -no-install flag, to run
against installed libflux-core.so or other libflux-* libraries.
This not only checks the incorrect code, but causes hard failures
in unit tests when the "poison" libflux-* libraries are installed
in CI.

Reverse the order that libflux-internal.la and libflux-core.la are
listed in LDADD for unit and other test executables. This appears to
resolve the issue, though the reason is unknown.