sharness: prepend src/common/.libs to LD_LIBRARY_PATH
Problem: There are some programs in the testsuite for which libtool
always adds an -rpath to the system directory, and thus these programs
may pick up system versions of libflux-*.so instead of the build
tree versions. Best case, the wrong code is tested, but also test
failures and crashes can occur. This is particularly a problem with
the MPI tests, since it appears that linking with the MPI libraries is
something that triggers libtool to add an RPATH to system library path.

Since we know we want to always link with the built libflux-*.so
libraries (except when TEST_INSTALL is active), always prepend
${abs_top_builddir}/src/common/.libs to LD_LIBRARY_PATH for sharness
tests if testing in the builddir. This should work around any cases
where libtool adds /usr/lib64 or other system path to -rpath.