broker/overlay: add overlay_peer_is_torpid()
Problem: the broker logs a message at LOG_ERR level when no messages,
including heartbeat-timed keepalives, have been received from a broker
for a period of time, but there is no way for another part of the broker
to test whether a peer is unresponsive.

Add an accessor to test whether a peer is "torpid" (unresponsive).
Rename internal variables and functions to use the term "torpid"
instead of "idle" which is a bit overloaded.