broker/overlay: invoke monitor on torpid change
Problem: there is no way to monitor broker peer unresponsiveness.

Invoke the overlay monitor callback when a TBON child becomes torpid,
or when a torpid child is finally heard from.

The callback may use the overlay_peer_is_torpid() function to test whether
the callback rank is torpid or not.