broker/groups: track broker.torpid group
Problem: there is no way outside of the broker to monitor peer
responsiveness, but the resource module needs to automatically
drain unresponsive nodes.

Add the 'broker.torpid' group, which will reflect the current set of
broker ranks that are deemed torpid by their TBON parent.  The
'groups.get' RPC may be used to monitor changes in the group.

Brokers that transition to 'offline' or 'lost' state are automatically
removed from the 'broker.torpid' group.

Caveat: in the case of a router node, only the router node is marked
torpid, not its entire subtree, even though the subtree is only reachable
through the torpid router node.