python: uri: add fallback for getting broker child pid
Problem: Some systems may not support /proc/<pid>/task/<tid>/children,
but it is required that the pid:<pid> URI resolver be able to find a
child if <pid> refers to a flux-broker.

Add a fallback method (still Linux-only) that tries to find a child
of PID by reading all /proc/<pid>/stat until a PPID that matches
the broker PID is found.

Determine if the fallback is necessary at initialization of the PID
URIResolver class, which should be more efficient when multiple URIs
are resolved, and also allows an environment variable to be supported
for testing purposes.

If FLUX_FORCE_BROKER_CHILD_FALLBACK is set, then force use of the
fallback method, since all distros used in ci support
/proc/<pid>/task/<tid>/children.

Fixes #4049