flux-perilog-run: improve error logging
Problem: When `flux perilog-run --exec-per-rank=CMD` fails, error
messages from individual ranks are not differentiated, so it is
difficult to tell on which rank/host the prolog or epilog error
occurred. Additionally, the default Python logging level is too low
and important "info" messages are dropped.

Adjust the execution of subprocesses so that stderr is captured,
and log any stderr from failed ranks. Also capture the instance
hostlist before running and use this to display the failing hostname
along with its rank for easier diagnosis.

Finally, promote the log message regarding draining nodes from
LOGGER.info to LOGGER.error, so that it is displayed by default.

Fixes #4053