doc: split out manpages list into own py file
Problem: The generation of commands to be listed in
'flux help' unnecessarily requires the Python docutils package
because the manpages list is stored in a python file that
imports docutils.

Solution: Split out the manpages list into its own file, removing
the docutils requirement for 'flux help' to be generated properly.

Fixes #4055