flux-proxy: attempt to restore terminal on error
Problem: When flux-proxy(1) is being used interactively and the Flux
instance to which the proxy connection is established terminates,
the current terminal can be left in a bad state.

Attempt to restore original terminal settings when flux_reactor_run(3)
exits with an error. Do nothing if stdin is not a tty.

The shell which was spawned from flux-proxy(1) may still print "exit"
when the user types the next character, but restoring the terminal
setting should help with issues like tty left in raw state or `echo`
disabled.

Fixes #2193