flux-proxy: always wait for child to exit, add --kill option
Problem: flux-proxy(1) exits immediately on error, leaving
any remaining child process (e.g. an interactive shell) to be
reparented. This causes the child process to become unusable in most
cases, which may be unfriendly when a user is doing something in that
child process.

When encountering an error in flux_reactor_run(3), always wait for
the child process to exit. Print a message to this effect to give a
hint to the user that flux-proxy(1) is waiting for any child to exit.

Add a new -k, --kill option to have flux-proxy(1) optionally send
SIGTERM to the child process on error.