flux-mini: change pty option to pty.interactive
Problem: The default behavior of the job shell `pty` option will
 be changing, which will break usage of `flux mini alloc`.

Change the pty option set by default in `flux mini alloc` to be
`pty.interactive` instead of just `pty`. This is the planned option
for allocating a pty to rank 0 for interactive use, but it won't
break the current use case since the shell "pty" option is still set.