libterminus: add flux_pty_monitor()
Problem: There is no way to locally monitor data being sent through a
`struct flux_pty`, e.g. in case the data is to be logged or otherwise
captured. Instead, a client would have to be attached to the pty,
which would require sending messages to ourselves, a paradigm not
easily supported in many simple scenarios.

Add a monitor callback to struct flux_pty, which can be set via a new
flux_pty_monitor() callback. If set, this callback will be invoked
when any new data is read from the pty.

Instead of adding a data/arg pointer that is passed to the
callback, extend the flux_pty interface with flux_pty_aux_set()
and flux_pty_aux_get(), so that users may stash arbitrary data in
the flux_pty for use in the monitor callback.