shell: subscribe to shell.output events in output plugin
Problem: It may be useful to copy non-task output to the same location
that task stdout and/or stderr are redirected from job shell plugins
or other shell entities, but the shell does not provide an easy way
to do that.

Subscribe to the "shell.output" callback topic in the output plugin,
which expects an ARG_IN argument of the form

 { "stream":s, "rank":i, "data":s% }

which will then be handled directly by shell_output_write(), so the
data will appear to have come from the designated task and will end
up in the same final location (e.g. kvs, file, etc).