t2500-job-attach.t: change -o pty to -o pty.interactive
Problem: The default behavior for the shell pty option will change,
which may break the pty test in t2500-job-attach.t.

Update the test to use '-o pty.interactive' and indicate the test is
related to interactive pty attach. Since the pty option is still set,
the test will not break with current code.