shell: pty: expand job shell pty support
Problem: The pty support in the job shell is currently limited. It only
supports creating a pty for task rank 0, and by default output produced
under a pty is not captured anywhere. This is fine for interactive
use, as with `flux mini alloc`, but does not cover use cases where
a pty is required non-interactively (e.g. for color output), or is
needed on all tasks.

Update the job shell pty option to allow an object to give more
control over how and where ptys are allocated across a job. The
allowed members in the new pty object are:

 {
    ranks:{s or i}   # rank or ranks to run under a pty, "all" for
                     #  all ranks

    capture:i        # if nonzero, capture pty output to same
                     #  destination as stdout

    interactive:i    # Setup for "interactive" attach by logging the
                     #  service endpoint for task rank 0 pty along with
                     #  the shell.init event.
 }

All of the above members are optional. The default if the pty option
is not an object, e.g. `-o pty`, is to allocate ptys for all ranks,
capturing data non-interactively, i.e.

 { "ranks":"all", "capture":1, "interactive":0 }

Therefore, this is a change from current behavior.

If only "interactive" is specified, then "ranks" defaults to 0, and
"capture" is disables. Therefore, previous behavior of the pty option
can be obtained with `-o pty.interactive`.

To get an interactive pty but still capture output to the KVS, e.g.
for later perusal, simply add the pty.capture option after the
interactive option, e.g.

 $ flux mini run -o pty.interactive -o pty.capture ...