libflux: add flux_get_instance_starttime()
Problem: the broker.starttime attribute is only accurate as a
stand-in for instance start time if fetched from rank 0, but the
attribute functions do not allow a rank to be specified.

Add flux_get_instance_starttime() convenience function.
With this, users won't need to set up the attr.get RPC directly.

Update unit test.