flux-pstree: display runtime of non-job instance
Problem: When pstree is displaying extra details of an instance that
is not a job, it cannot display runtime since it cannot query the parent
for the `t_run` attribute, so this field is annoyingly blank.

Now that the broker.starttime attribute is supported, attempt to fetch
its value when the parent cannot be queried. This allows a runtime to
be displayed.

Fixes #4033