flux-mini: allow exec events to be used with --wait-event
Problem: A calling program may want to cause submit/bulksumit to block
until a job gets one of the exec events, like shell.init, but waiting on
events in the exec eventlog is not currently supported.

Support waiting on exec events by prepending the --wait-event argument with
"exec.", e.g. `--wait-event=exec.shell.init`.

If args.verbose > 2, log exec events in addition to normal events,
but only if the exec eventlog is being watched due to --watch or
--wait-event=exec.*.