ci: rename centos builders to "el"
Problem: CentOS reached end-of-life Dec 31, 2021, and so yum repos can
no longer be accessed. We'd like to replace the centos8 image with
a different RHEL8 clone distro, but this will require updating all
references to 'centos8' to another specific distro and version number.

Prepare for transition to different RHEL clone images by renaming
centos to a more generic name "el" globally. This will allow the
transition to different distros to occur within the Dockerfile(s)
only, so that global renaming won't be necessary next time around.

Use of the generic "el" name for Enterprise Linux indicates that
these builders are generically Red Hat Enterprise Linux clones, and
thus allows us to update the underlying base images in the future
without the need for this same global renaming exercise.