shell: disable stderr logging when evlog is initialized
Problem: The shell emits duplicate log messages to both stderr and the
output eventlog for FLUX_SHELL_ERROR and below. This is because the
"evlog" plugin sets the stderr level to FLUX_SHELL_ERROR instead of
completely disabling stderr logging.

Disable stderr logging completely when the evlog plugin is initialized.
This removes the duplicate messages injected into the output eventlog
by job-exec when the shell emits errors on stderr.

The likelihood of "losing" important errors is low. The evlog plugin
will emit any dropped messages to stderr if they fail to be committed
to the output eventlog. Fatal errors additionally result in a job
exception, which typically contains the log message which caused the
fatal exception.