shell: evlog: use stderr logging after evlog shutdown
Problem: Log messages may be lost if they are posted after the
evlog plugin is torn down.

Reset the stderr logging level before destroying the evlog plugin.
This should allow log messages sent after the evlog plugin is destroyed
to be emitted to stderr and eventually end up in the job's output
eventlog.