flux-overlay: lookup hosts/ranks both ways
Problem: flux overlay gethostbyrank translates a rank idset
to a hostlist, but there is no similar subcommand to translate
a hostlist to a rank idset.

Rather than adding a separate subcommand, rename the gethostbyrank
subcommand to "lookup" and allow it to take either an idset or a
hostlist as input.

Note that in the case of multiple brokers per node, only the first
rank running on a target host is printed.

Update sharness test.

Fixes #4091