flux-overlay: support disconnect HOST
Problem: flux overlay disconnect fails to detect malformed rank
argument, and does not accept hostnames.

Add better error checking on rank parsing, and convert hostname
to rank if a valid hostname is specified.

Note that when there are multiple brokers per node, the hostname
is mapped to the first rank.

Add tests

Fixes #4090