job-manager: perilog: fix segfault on job-exception
Problem: The perilog jobtap plugin stashes a "perilog_proc" aux item
in the associated job using flux_jobtap_job_aux_set(), and then uses
the existince of this item to determine if there is still a perilog
process running when a job exception is raised. The plugin does not,
however, delete the "perilog_proc" aux item when the prolog/epilog
completes, so if the exception is raised after the prolog completes,
for example, then a segfault results due to use-after-free.

Clear the perilog_proc aux item for the associated job before freeing
the memory associated with a completed perilog process to avoid
this crash.

Fixes #4089