broker: use TCP level keepalives
Problem: when a node is turned off, Flux RPCs to or through that
node can go unanswered indefinitely.

When a node is abrubtly turned off, as opposed to a broker killed, the
node's kernel cannot send a FIN packet to close the TCP connection.
The connection remains active until the node is back online and can
respond with an RST packet to abort the connection.  The traditional
solution to this problem is to enable TCP level keepalives.

Enable TCP keepalives for TBON children.  After the connection goes idle
for some period of time, keepalive probes are sent, and if they are
unanswered, the kernel drops the connection.  A lost child connection is
already handled by the broker - pending RPCs receive EHOSTUNREACH responses.

The default tunables (see sysctl -a | grep tcp_keepalive) may be overridden
by config file or on the broker command line.  They cannot be changed on a
live Flux instance as zeromq requires these parameters to be set before
the socket is bound.  Keepalives may also be disabled via config file or
broker command line.

TCP keepalives complement the application level keepalives.  TCP keepalives
can elicit a response from the kernel even when the broker process is
unresponsive and cannot send application level keepalives, e.g. possibly
due to temporary conditions like low memory or high message load.  Flux
drains nodes with unresponsive brokers, but does not forcibily disconnect
them unless the kernel/network is also unresponsive.

This commit only enables TCP keepalives on downstream TBON children, not
on the parent.  Since normal operation is not possible until the parent is
back online, there is no advantage to early failure detection.

Fixes #4086